refactor: 사물함 모델 유틸 파일 분리 #53
