+++ apps/web/src/widgets/locker/model/index.ts
@@ -1,3 +1,5 @@
-export * from './lib';
+export * from './constants';
+export * from './format';
+export * from './lockerStatus';
 export * from './types';
 export * from './useLockerPage';
